Extract SetTagKeyMixin from mixin.rb
